Split PHPUnit tests
