+++ tests/Bancard_00_BasicTests.php
@@ -1,0 +1,49 @@
+<?php declare(strict_types=1);
+
+namespace HDSSolutions\Bancard\Tests;
+
+use HDSSolutions\Bancard\Bancard;
+use HDSSolutions\Bancard\Models\Currency;
+use PHPUnit\Framework\TestCase;
+use RuntimeException;
+
+final class Bancard_00_BasicTests extends TestCase {
+
+    public function testThatWeAreOnTestingEnvironment(): void {
+        $this->assertTrue(Bancard::isDevelop());
+    }
+
+    public function testThatCredentialsCanBeSet(): void {
+        Bancard::credentials(
+            publicKey: BANCARD_PUBLIC_KEY,
+            privateKey: BANCARD_PRIVATE_KEY,
+        );
+        $this->assertSame(Bancard::getPrivateKey(), BANCARD_PRIVATE_KEY);
+    }
+
+    public function testInvalidDataValidations(): void {
+        $this->expectException(RuntimeException::class);
+        Bancard::single_buy(
+            shop_process_id: 0,
+            amount:          0,
+            description:     '',
+            currency:        '??',
+            return_url:      '',
+            cancel_url:      '',
+        );
+    }
+
+    public function testInvalidDataRequest(): void {
+        $single_buy = Bancard::single_buy(
+            shop_process_id: random_int(10000, 99999),
+            amount:          -10,
+            description:     'Test invalid amount',
+            currency:        Currency::Dollar,
+            return_url:      '',
+            cancel_url:      '',
+        );
+        $this->assertFalse($single_buy->wasSuccess());
+        $this->assertNotEmpty($single_buy->getMessages());
+    }
+
+}
